backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
aspose-words==24.3.0
Babel==2.14.0
black==23.11.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
docx2pdf==0.1.8
docxcompose==1.4.0
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
isort==5.12.0
lxml==5.2.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
psutil==5.9.8
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
